pyproject.toml
[project]
name = "appium-tools"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "appium-python-client>=5.2.4",
    "langchain>=1.0.8",
    "langchain-community>=0.4.1",
    "langchain-core>=1.1.0",
    "langchain-openai>=1.0.3",
    "langgraph>=1.0.3",
    "openai>=2.8.1",
]
